Avoid hanging without timeout during the SSL handshake phase of an SSL connection.
See https://issues.apache.org/jira/browse/HTTPCLIENT-1478 for details
of the problem. The comments there, from 27/Nov/14, show that a fix
went into HttpClient 4.3.6, that consists of setting the SO_TIMEOUT
to a positive value. (They use the connectTimeout as a fallback.)

We add a custom socket configuration, in order to set the default
socket timeout, from the value that was passed to the
ConnectOptions.setSocketTimeout method.